avoid extra redundant copying for obj type spec info
