+++ lib/Backend/Backend.h
@@ -16,8 +16,6 @@
 #include "Language/AsmJsModule.h"
 #include "Language/ProfilingHelpers.h"
 #include "Language/FunctionCodeGenRuntimeData.h"
-#include "Language/ObjTypeSpecFldInfo.h"
-#include "Language/FunctionCodeGenJitTimeData.h"
 #include "Language/JavascriptMathOperators.h"
 #include "Language/JavascriptMathOperators.inl"
 #include "Language/JavascriptStackWalker.h"
@@ -112,7 +110,7 @@ enum IRDumpFlags
 
 #include "JITTimeProfileInfo.h"
 #include "JITRecyclableObject.h"
-#include "JITTimeFixedField.h"
+#include "FixedFieldInfo.h"
 #include "JITTimePolymorphicInlineCache.h"
 #include "JITTimePolymorphicInlineCacheInfo.h"
 #include "CodeGenWorkItemType.h"
@@ -120,7 +118,8 @@ enum IRDumpFlags
 #include "JITTimeConstructorCache.h"
 #include "JITTypeHandler.h"
 #include "JITType.h"
-#include "JITObjTypeSpecFldInfo.h"
+#include "ObjTypeSpecFldInfo.h"
+#include "FunctionCodeGenJitTimeData.h"
 #include "ServerScriptContext.h"
 #include "JITOutput.h"
 #include "JITTimeScriptContext.h"
