+++ lib/Backend/FunctionCodeGenJitTimeData.cpp
@@ -2,8 +2,8 @@
 // Copyright (C) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 //-------------------------------------------------------------------------------------------------------
-#include "RuntimeLanguagePch.h"
-#include "JITTimeFunctionBody.h"
+
+#include "Backend.h"
 
 #if ENABLE_NATIVE_CODEGEN
 namespace Js
