+++ lib/Backend/FunctionJITTimeInfo.h
@@ -35,8 +35,8 @@ public:
     const FunctionJITRuntimeInfo * GetRuntimeInfo() const;
     const BVFixed * GetInlineesBV() const;
     const FunctionJITTimeInfo * GetJitTimeDataFromFunctionInfoAddr(intptr_t polyFuncInfo) const;
-    JITObjTypeSpecFldInfo * GetObjTypeSpecFldInfo(uint index) const;
-    JITObjTypeSpecFldInfo * GetGlobalObjTypeSpecFldInfo(uint index) const;
+    ObjTypeSpecFldInfo * GetObjTypeSpecFldInfo(uint index) const;
+    ObjTypeSpecFldInfo * GetGlobalObjTypeSpecFldInfo(uint index) const;
     uint GetGlobalObjTypeSpecFldInfoCount() const;
     const FunctionJITRuntimeInfo * GetInlineeForTargetInlineeRuntimeData(const Js::ProfileId profiledCallSiteId, intptr_t inlineeFuncBodyAddr) const;
     const FunctionJITRuntimeInfo *GetInlineeRuntimeData(const Js::ProfileId profiledCallSiteId) const;
